Tuesday upload
Made the files on Monday, but now moved them into a repository and made some of the functionality.
Have made buttons for displaying more or less results per page, which affects a value, but there's no call for this value beyond the one that runs when the page is refreshed.